「管理者チェックを追加」

+++ src/main/java/actions/EmployeeAction.java
@@ -42,27 +42,32 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException, IOException {
 
-        //指定されたページ数の一覧画面に表示するデータを取得
-        int page = getPage();
-        List<EmployeeView> employees = service.getPerPage(page);
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        //全ての従業員データの件数を取得
-        long employeeCount = service.countAll();
+            //指定されたページ数の一覧画面に表示するデータを取得
+            int page = getPage();
+            List<EmployeeView> employees = service.getPerPage(page);
 
-        putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
-        putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
-        putRequestScope(AttributeConst.PAGE, page); //ページ数
-        putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+            //全ての従業員データの件数を取得
+            long employeeCount = service.countAll();
 
-        //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
-        String flush = getSessionScope(AttributeConst.FLUSH);
-        if (flush != null) {
-            putRequestScope(AttributeConst.FLUSH, flush);
-            removeSessionScope(AttributeConst.FLUSH);
+            putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
+            putRequestScope(AttributeConst.PAGE, page); //ページ数
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
+
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
+
         }
-
-        //一覧画面を表示
-        forward(ForwardConst.FW_EMP_INDEX);
 
     }
 
@@ -71,19 +76,23 @@ public class EmployeeAction extends ActionBase {
      * @throws ServletException
      * @throws IOException
      */
-    public void entryNew() throws ServletException, IOException{
+    public void entryNew() throws ServletException, IOException {
 
-        putRequestScope(AttributeConst.TOKEN,getTokenId()); //CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
-        //新規登録画面を表示
-        forward(ForwardConst.FW_EMP_NEW);
+            //新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        }
+
     }
 
-    public void create() throws ServletException,IOException{
+    public void create() throws ServletException, IOException {
 
         //CSRF対策 tokenチェック
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             //パラメータの値をもとに従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -102,24 +111,24 @@ public class EmployeeAction extends ActionBase {
             //従業員情報取得
             List<String> errors = service.create(ev, pepper);
 
-            if(errors.size() >0) {
+            if (errors.size() > 0) {
                 //登録中にエラーがあった場合
 
-                putRequestScope(AttributeConst.TOKEN,getTokenId()); //CSRF対策用トークン
-                putRequestScope(AttributeConst.EMPLOYEE,ev); //入力された従業員情報
+                putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+                putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
                 putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
 
                 //新規登録画面を再表示
                 forward(ForwardConst.FW_EMP_NEW);
 
-            }else {
+            } else {
                 //登録中にエラーがなかった場合
 
                 //セッションに登録完了のフラッシュメッセージを設定
                 putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
 
                 //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
 
             }
 
@@ -131,22 +140,27 @@ public class EmployeeAction extends ActionBase {
      * @throws ServletException
      * @throws IOException
      */
-    public void show() throws ServletException, IOException{
+    public void show() throws ServletException, IOException {
 
-        //idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //詳細画面を表示
+            forward(ForwardConst.FW_EMP_SHOW);
         }
 
-        putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
-
-        //詳細画面を表示
-        forward(ForwardConst.FW_EMP_SHOW);
     }
 
     /**
@@ -154,23 +168,28 @@ public class EmployeeAction extends ActionBase {
      * @throws ServletException
      * @throws IOException
      */
-    public void edit() throws ServletException, IOException{
+    public void edit() throws ServletException, IOException {
 
-        //idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
 
-        if(ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策トークン
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //編集画面を表示する
+            forward(ForwardConst.FW_EMP_EDIT);
         }
 
-        putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策トークン
-        putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
-
-        //編集画面を表示する
-        forward(ForwardConst.FW_EMP_EDIT);
     }
 
     /**
@@ -178,10 +197,10 @@ public class EmployeeAction extends ActionBase {
      * @throws ServletException
      * @throws IOException
      */
-    public void update() throws ServletException, IOException{
+    public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) {
             //パスワードの値を元に従業員のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -197,9 +216,9 @@ public class EmployeeAction extends ActionBase {
             String pepper = getContextScope(PropertyConst.PEPPER);
 
             //従業員情報更新
-            List<String> errors = service.update(ev,  pepper);
+            List<String> errors = service.update(ev, pepper);
 
-            if(errors.size() > 0) {
+            if (errors.size() > 0) {
                 //更新中にエラーが発生した場合
 
                 putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
@@ -208,7 +227,7 @@ public class EmployeeAction extends ActionBase {
 
                 //編集画面を再表示
                 forward(ForwardConst.FW_EMP_EDIT);
-            }else {
+            } else {
                 //更新中にエラーがなかった場合
 
                 //セッションに更新完了のフラッシュメッセージを設定
@@ -220,24 +239,48 @@ public class EmployeeAction extends ActionBase {
         }
 
     }
+
     /**
      * 論理削除を行う
      * @throws ServretException
      * @throws IOException
      */
-    public void destroy() throws ServletException, IOException{
+    public void destroy() throws ServletException, IOException {
 
         //CSRF対策tokenのチェック
-        if(checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
             //セッションに削除完了のフラッシュメッセージを設定
-            putSessionScope(AttributeConst.FLUSH,MessageConst.I_DELETED.getMessage());
+            putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
 
             //一覧画面にリダイレクト
-            redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+            redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+
+            return true;
         }
     }
 }
